+++ CHANGELOG.md
@@ -8,6 +8,7 @@ v2.3.x
 - Merge PR from @MacKey-255 with new enhancements
 - Send and wait APIs (```SendAndWaitAsync```)
 - Client connect with timeout API (```StartWithTimeout``` and ```StartWithTimeoutAsync```)
+- Multiple listener prefixes
 
 ## Previous Versions